split out agent factory
