+++ service/src/services/matrix-admin/rooms/matrix.admin.rooms.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MatrixAdminRoomsService } from './matrix.admin.rooms.service';
 import { CommunicationAdapterModule } from '../../communication-adapter/communication-adapter.module';
-import { MatrixAgentModule } from '../../matrix/agent/matrix.agent.module';
+import { MatrixAgentModule } from '../../matrix/agent-factory/matrix.agent.factory.module';
 import { MatrixUserAdapterModule } from '../../matrix/adapter-user/matrix.user.adapter.module';
 import { MatrixAdminRoomsController } from './matrix.admin.rooms.controller';
 import { MatrixAdminUserElevatedModule } from '../user-elevated/matrix.admin.user.elevated.module';
