+++ service/src/services/matrix-admin/user-elevated/matrix.admin.user.elevated.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MatrixAgentModule } from '@services/matrix/agent/matrix.agent.module';
+import { MatrixAgentModule } from '@src/services/matrix/agent-factory/matrix.agent.factory.module';
 import { MatrixUserAdapterModule } from '@services/matrix/adapter-user/matrix.user.adapter.module';
 import { MatrixAdminUserElevatedService } from './matrix.admin.user.elevated.service';
 import { MatrixUserManagementModule } from '../user/matrix.admin.user.module';
