+++ service/src/services/matrix-admin/user-elevated/matrix.admin.user.elevated.service.ts
@@ -6,7 +6,7 @@ import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { MatrixUserAdapter } from '@services/matrix/adapter-user/matrix.user.adapter';
 import { IOperationalMatrixUser } from '@services/matrix/adapter-user/matrix.user.interface';
 import { MatrixAgent } from '@services/matrix/agent/matrix.agent';
-import { MatrixAgentService } from '@services/matrix/agent/matrix.agent.service';
+import { MatrixAgentFactoryService } from '@src/services/matrix/agent-factory/matrix.agent.factory.service';
 import { MatrixUserManagementService } from '../user/matrix.admin.user.service';
 
 @Injectable()
@@ -20,7 +20,7 @@ export class MatrixAdminUserElevatedService {
   constructor(
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
     private readonly logger: pkg.LoggerService,
-    private matrixAgentService: MatrixAgentService,
+    private matrixAgentService: MatrixAgentFactoryService,
     private configService: ConfigService,
     private matrixUserManagementService: MatrixUserManagementService,
     private matrixUserAdapter: MatrixUserAdapter
