+++ service/src/services/matrix/agent-factory/matrix.agent.factory.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { MatrixMessageAdapterModule } from '../adapter-message/matrix.message.adapter.module';
 import { MatrixRoomAdapterModule } from '../adapter-room/matrix.room.adapter.module';
 import { MatrixUserAdapterModule } from '../adapter-user/matrix.user.adapter.module';
-import { MatrixAgentService } from './matrix.agent.service';
 import { MatrixUserManagementModule } from '@src/services/matrix-admin/user/matrix.admin.user.module';
+import { MatrixAgentFactoryService } from './matrix.agent.factory.service';
 
 @Module({
   imports: [
@@ -12,7 +12,7 @@ import { MatrixUserManagementModule } from '@src/services/matrix-admin/user/matr
     MatrixRoomAdapterModule,
     MatrixMessageAdapterModule,
   ],
-  providers: [MatrixAgentService],
-  exports: [MatrixAgentService],
+  providers: [MatrixAgentFactoryService],
+  exports: [MatrixAgentFactoryService],
 })
 export class MatrixAgentModule {}
