+++ service/src/services/matrix/agent-pool/matrix.agent.pool.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { MatrixAgentPool } from '@services/matrix/agent-pool/matrix.agent.pool';
-import { MatrixAgentModule } from '../agent/matrix.agent.module';
+import { MatrixAgentModule } from '../agent-factory/matrix.agent.factory.module';
 import { MatrixUserManagementModule } from '@src/services/matrix-admin/user/matrix.admin.user.module';
 
 @Module({
