+++ service/src/services/matrix/agent-pool/matrix.agent.pool.ts
@@ -5,7 +5,7 @@ const { Inject, Injectable } = pkg;
 import { OnModuleDestroy, OnModuleInit } from '@nestjs/common';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { MatrixAgent } from '../agent/matrix.agent';
-import { MatrixAgentService } from '../agent/matrix.agent.service';
+import { MatrixAgentFactoryService } from '../agent-factory/matrix.agent.factory.service';
 import { ConfigService } from '@nestjs/config';
 import { Disposable } from '@src/common/interfaces/disposable.interface';
 import { NotSupportedException } from '@src/common/exceptions/not.supported.exception';
@@ -21,7 +21,7 @@ export class MatrixAgentPool
   private _agentPoolSize: number;
 
   constructor(
-    private matrixAgentService: MatrixAgentService,
+    private matrixAgentService: MatrixAgentFactoryService,
     private configService: ConfigService,
     private matrixUserManagementService: MatrixUserManagementService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
